+++ src/app/components/embedded-content-private-share.js
@@ -7,26 +7,19 @@
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 
-import OneEmbeddedContainer from 'onezone-gui/components/one-embedded-container';
+import OneproviderEmbeddedContainer from 'onezone-gui/components/oneprovider-embedded-container';
 import layout from 'onezone-gui/templates/components/one-embedded-container';
 import { inject as service } from '@ember/service';
-import { computed } from '@ember/object';
 import { reads } from '@ember/object/computed';
 import EmbeddedContentShareActions from 'onezone-gui/mixins/embedded-content-share-actions';
 
-export default OneEmbeddedContainer.extend(EmbeddedContentShareActions, {
+export default OneproviderEmbeddedContainer.extend(EmbeddedContentShareActions, {
   layout,
 
   navigationState: service(),
   globalNotify: service(),
   shareManager: service(),
   router: service(),
-
-  /**
-   * @virtual
-   * @type {Models.Provider}
-   */
-  oneprovider: undefined,
 
   /**
    * @virtual
@@ -65,11 +58,6 @@ export default OneEmbeddedContainer.extend(EmbeddedContentShareActions, {
   /**
    * @override implements OneEmbeddedContainer
    */
-  iframeType: 'oneprovider',
-
-  /**
-   * @override implements OneEmbeddedContainer
-   */
   iframeInjectedProperties: Object.freeze([
     'shareId',
     'spaceId',
@@ -85,19 +73,6 @@ export default OneEmbeddedContainer.extend(EmbeddedContentShareActions, {
     'showSharesIndex',
     'reloadShareList',
   ]),
-
-  /**
-   * @override implements OneEmbeddedContainer
-   */
-  relatedData: reads('oneprovider'),
-
-  /**
-   * @override implements OneEmbeddedContainer
-   */
-  iframeId: computed('oneprovider.entityId', function iframeId() {
-    const oneproviderId = this.get('oneprovider.entityId');
-    return `iframe-oneprovider-${oneproviderId}`;
-  }),
 
   actions: {
     showSharesIndex() {