+++ src/app/components/embedded-content-space-transfers.js
@@ -7,24 +7,16 @@
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 
-import OneEmbeddedContainer from 'onezone-gui/components/one-embedded-container';
+import OneproviderEmbeddedContainer from 'onezone-gui/components/oneprovider-embedded-container';
 import layout from 'onezone-gui/templates/components/one-embedded-container';
 import { inject as service } from '@ember/service';
-import { computed } from '@ember/object';
-import { reads } from '@ember/object/computed';
 
-export default OneEmbeddedContainer.extend({
+export default OneproviderEmbeddedContainer.extend({
   layout,
 
   navigationState: service(),
   globalNotify: service(),
   router: service(),
-
-  /**
-   * @virtual
-   * @type {Models.Provider}
-   */
-  oneprovider: undefined,
 
   /**
    * Entity ID of `space` record that is space of directory displayed in files
@@ -55,11 +47,6 @@ export default OneEmbeddedContainer.extend({
   /**
    * @override implements OneEmbeddedContainer
    */
-  iframeType: 'oneprovider',
-
-  /**
-   * @override implements OneEmbeddedContainer
-   */
   iframeInjectedProperties: Object.freeze([
     'spaceEntityId',
     'fileEntityId',
@@ -74,19 +61,6 @@ export default OneEmbeddedContainer.extend({
     'resetQueryParams',
     'changeListTab',
   ]),
-
-  /**
-   * @override implements OneEmbeddedContainer
-   */
-  relatedData: reads('oneprovider'),
-
-  /**
-   * @override implements OneEmbeddedContainer
-   */
-  iframeId: computed('oneprovider.entityId', function iframeId() {
-    const oneproviderId = this.get('oneprovider.entityId');
-    return `iframe-oneprovider-${oneproviderId}`;
-  }),
 
   actions: {
     closeFileTab() {